MVC refactor step #1 - new package layout
