+++ app/blueprint.py
@@ -56,7 +56,7 @@ class Blueprint(flask.Blueprint):
         """
         if self.__index_route is not None:
             endpoint, f, options = self.__index_route
-            if parse_pkg(app.modules["main"]["module"])[1] == self.name:
+            if parse_pkg(app.modules["main"]["module"])[-1] == self.name:
                 # this will be the true index of our site!
                 self.add_url_rule("/", endpoint, f, **options)
                 self.is_registered_index = True
