+++ app/controllers/main/forms.py
@@ -7,7 +7,7 @@ from wtforms import (
     ValidationError,
 )
 from wtforms.validators import DataRequired, Length, Email, EqualTo
-from .models import User
+from tassaron_flask_template.models.main.models import User
 
 
 class LoginForm(FlaskForm):
