+++ app/controllers/main/task_overview.py
@@ -7,8 +7,8 @@ from flask import render_template
 blueprint = Blueprint(
     "huey",
     __name__,
-    static_folder="../static",
-    template_folder="../templates",
+    static_folder="../../static",
+    template_folder="../../templates",
 )
