+++ app/controllers/shop/cart_api.py
@@ -3,7 +3,7 @@ POST to these Cart endpoints in order to manipulate the Cart Session Cookie
 Each endpoint returns a response of success or not, to update the client-side record
 """
 from flask import Blueprint, request, session
-from tassaron_flask_template.main.plugins import db
+from tassaron_flask_template.models.main.plugins import db
 from .inventory_models import Product
