+++ app/controllers/shop/inventory.py
@@ -2,13 +2,13 @@ from flask import current_app, render_template, flash, redirect, url_for, abort
 from flask_login import current_user
 from werkzeug.datastructures import MultiDict
 from tassaron_flask_template.blueprint import Blueprint
-from tassaron_flask_template.main.plugins import db
+from tassaron_flask_template.models.main.plugins import db
 from .inventory_forms import ProductForm
 from .inventory_models import Product
 import os
 
 
-blueprint = Blueprint("inventory", __name__, template_folder="../templates/inventory")
+blueprint = Blueprint("inventory", __name__, template_folder="../../templates/inventory")
 
 
 @blueprint.admin_route("")
