+++ app/controllers/shop/inventory_forms.py
