+++ tests/test_cart_api.py
@@ -1,6 +1,6 @@
-from tassaron_flask_template.main import create_app, init_app
-from tassaron_flask_template.main.plugins import db
-from tassaron_flask_template.shop.inventory_models import Product, ProductCategory
+from tassaron_flask_template.models.main import create_app, init_app
+from tassaron_flask_template.models.main.plugins import db
+from tassaron_flask_template.controllers.shop.inventory_models import Product, ProductCategory
 import tempfile
 import os
 from flask import json, jsonify
