+++ tests/test_email_verification.py
@@ -1,6 +1,6 @@
-from tassaron_flask_template.main import create_app, init_app
-from tassaron_flask_template.main.plugins import db
-from tassaron_flask_template.main.models import User
+from tassaron_flask_template.models.main import create_app, init_app
+from tassaron_flask_template.models.main.plugins import db
+from tassaron_flask_template.models.main.models import User
 from tassaron_flask_template.email import *
 from huey.api import Result
 import tempfile
