+++ tests/test_profile.py
@@ -3,9 +3,9 @@ import tempfile
 import pytest
 import flask_login
 from flask import url_for
-from tassaron_flask_template.main import create_app, init_app
-from tassaron_flask_template.main.plugins import db, bcrypt, login_manager
-from tassaron_flask_template.main.models import User, ShippingAddress
+from tassaron_flask_template.models.main import create_app, init_app
+from tassaron_flask_template.models.main.plugins import db, bcrypt, login_manager
+from tassaron_flask_template.models.main.models import User, ShippingAddress
 
 
 @pytest.fixture
