+++ tests/test_sessions.py
@@ -2,10 +2,10 @@ import os
 import tempfile
 import pytest
 import json
-from tassaron_flask_template.main import create_app, init_app
-from tassaron_flask_template.main.plugins import db
-from tassaron_flask_template.main.models import User
-from tassaron_flask_template.shop.inventory_models import *
+from tassaron_flask_template.models.main import create_app, init_app
+from tassaron_flask_template.models.main.plugins import db
+from tassaron_flask_template.models.main.models import User
+from tassaron_flask_template.controllers.shop.inventory_models import *
 from flask import session
